requirements.txt
fastapi
uvicorn
gunicorn
python-dotenv
pydantic
google-genai